// Frontend/src/Types/Confirmation.d.ts
// Generated by https://quicktype.io

export interface BookingInfo {
   iternaryInfo: IternaryInfo;
   paymentInfo: PaymentInfo;
   billingInfo: BillingInfo;
   travellerInfo: TravellerInfo;
}

export interface BillingInfo {
   billingInfoId: number;
   billingFirstName: string;
   billingLastName: string;
   billingAddress1: string;
   billingAddress2: string;
   billingCountry: string;
   billingCity: string;
   billingState: string;
   billingZip: string;
   billingPhone: string;
   billingEmail: string;
   booking: Booking;
}

export interface Booking {
   bookingId: number;
   bookingEmail: string;
   availabilities: number[];
   canceled: boolean;
}

export interface IternaryInfo {
   iternaryInfoId: number;
   checkInDate: string;
   checkOutDate: string;
   packageName: string;
   packageDescription: string;
   roomName: string;
   guestString: string;
   nightlyRate: number;
   subtotal: number;
   taxes: number;
   vat: number;
   total: number;
   booking: Booking;
   roomImage : string;
}

export interface PaymentInfo {
   paymentInfoId: number;
   cardNumber: string;
   expiryMonth: number;
   expiryYear: number;
   booking: Booking;
}

export interface TravellerInfo {
   travellerInfoId: number;
   firstName: string;
   lastName: string;
   phone: string;
   email: string;
   booking: Booking;
}